Refactor: move notification models to notifications/models
