+++ app/src/main/java/fr/free/nrw/commons/notification/NotificatinAdapter.kt
@@ -1,6 +1,6 @@
 package fr.free.nrw.commons.notification
 
-import fr.free.nrw.commons.data.models.notification.Notification
+import fr.free.nrw.commons.notification.models.Notification
 import fr.free.nrw.commons.upload.categories.BaseDelegateAdapter
 
 internal class NotificatinAdapter(onNotificationClicked: (Notification) -> Unit) :
