+++ app/src/main/java/fr/free/nrw/commons/notification/NotificationActivity.java
@@ -23,7 +23,7 @@ import butterknife.ButterKnife;
 import com.google.android.material.snackbar.Snackbar;
 import fr.free.nrw.commons.R;
 import fr.free.nrw.commons.Utils;
-import fr.free.nrw.commons.data.models.notification.Notification;
+import fr.free.nrw.commons.notification.models.Notification;
 import fr.free.nrw.commons.theme.BaseActivity;
 import fr.free.nrw.commons.utils.NetworkUtils;
 import fr.free.nrw.commons.utils.ViewUtil;
