+++ app/src/main/java/fr/free/nrw/commons/notification/NotificationController.java
@@ -1,6 +1,6 @@
 package fr.free.nrw.commons.notification;
 
-import fr.free.nrw.commons.data.models.notification.Notification;
+import fr.free.nrw.commons.notification.models.Notification;
 import java.util.List;
 
 import javax.inject.Inject;
