+++ app/src/main/java/fr/free/nrw/commons/notification/NotificationWorkerFragment.java
@@ -5,7 +5,7 @@ import android.os.Bundle;
 
 import androidx.annotation.Nullable;
 
-import fr.free.nrw.commons.data.models.notification.Notification;
+import fr.free.nrw.commons.notification.models.Notification;
 import java.util.List;
 
 /**
