+++ app/src/main/java/fr/free/nrw/commons/notification/models/Notification.kt
@@ -1,4 +1,4 @@
-package fr.free.nrw.commons.data.models.notification
+package fr.free.nrw.commons.notification.models
 
 import org.wikipedia.util.DateUtil
