+++ app/src/test/kotlin/fr/free/nrw/commons/notification/NotificationControllerTest.kt
@@ -1,6 +1,6 @@
 package fr.free.nrw.commons.notification
 
-import fr.free.nrw.commons.data.models.notification.Notification
+import fr.free.nrw.commons.notification.models.Notification
 import org.junit.Before
 import org.junit.Test
 import org.mockito.ArgumentMatchers
