+++ app/src/test/kotlin/fr/free/nrw/commons/notification/NotificationHelperUnitTests.kt
@@ -4,7 +4,8 @@ import android.app.NotificationManager
 import android.content.Context
 import android.content.Intent
 import fr.free.nrw.commons.TestCommonsApplication
-import fr.free.nrw.commons.data.models.notification.Notification
+import fr.free.nrw.commons.notification.models.Notification
+import fr.free.nrw.commons.notification.models.NotificationType
 import org.junit.Assert
 import org.junit.Before
 import org.junit.Test
